Changes requested on code review: * Unit tests moved into a dedicated specs class * MemberData moved directly below the test method that needs it * Use private methods instead of local functions
